backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Background Tasks
celery==5.3.4

# Validation and Serialization
pydantic==2.5.0

# QR Code Generation
qrcode[pil]==7.4.2

# Cloud Storage
cloudinary==1.36.0

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# HTTP Client
httpx==0.25.2

# File Upload and Storage
python-multipart==0.0.6
pillow==10.1.0

# Email
fastapi-mail==1.4.1

# CORS (built into FastAPI)
# fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0